# Optimize existAt calls on scavenged chunks using bloom filters for faster index merges