--- a/src/EventStore.Core.Tests/EventStore.Core.Tests.csproj
+++ b/src/EventStore.Core.Tests/EventStore.Core.Tests.csproj
@@ -654,6 +654,8 @@
     <Compile Include="Services\Replication\CommitReplication\with_index_committer_service.cs" />
     <Compile Include="Services\Replication\CommitReplication\when_single_node_cluster_receives_commit_ack_for_multiple_prepares.cs" />
     <Compile Include="Services\Replication\CommitReplication\when_constructing_index_committer_service.cs" />
+    <Compile Include="DataStructures\bloom_filter_should.cs" />
+    <Compile Include="TransactionLog\Optimization\tfchunkreader_existsat_optimizer_should.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\EventStore.ClientAPI.Embedded\EventStore.ClientAPI.Embedded.csproj">