--- a/src/EventStore.Core.Tests/TransactionLog/when_creating_chunked_transaction_chaser.cs
+++ b/src/EventStore.Core.Tests/TransactionLog/when_creating_chunked_transaction_chaser.cs
@@ -15,21 +15,21 @@ namespace EventStore.Core.Tests.TransactionLog
         public void a_null_file_config_throws_argument_null_exception()
         {
             Assert.Throws<ArgumentNullException>(
-                () => new TFChunkChaser(null, new InMemoryCheckpoint(0), new InMemoryCheckpoint(0)));
+                () => new TFChunkChaser(null, new InMemoryCheckpoint(0), new InMemoryCheckpoint(0), false));
         }
 
         [Test]
         public void a_null_writer_checksum_throws_argument_null_exception()
         {
             var db = new TFChunkDb(TFChunkHelper.CreateDbConfig(PathName, 0));
-            Assert.Throws<ArgumentNullException>(() => new TFChunkChaser(db, null, new InMemoryCheckpoint()));
+            Assert.Throws<ArgumentNullException>(() => new TFChunkChaser(db, null, new InMemoryCheckpoint(), false));
         }
 
         [Test]
         public void a_null_chaser_checksum_throws_argument_null_exception()
         {
             var db = new TFChunkDb(TFChunkHelper.CreateDbConfig(PathName, 0));
-            Assert.Throws<ArgumentNullException>(() => new TFChunkChaser(db, new InMemoryCheckpoint(), null));
+            Assert.Throws<ArgumentNullException>(() => new TFChunkChaser(db, new InMemoryCheckpoint(), null, false));
         }
     }
 }