… src/EventStore.Core/DataStructures/LRUCache.cs
… src/EventStore.Core/TransactionLog/Chunks/TFChunk/TFChunk.cs
… src/EventStore.Core/TransactionLog/Chunks/TFChunk/TFChunkReadSide.cs
… src/EventStore.Core/TransactionLog/Chunks/TFChunkDb.cs
… src/EventStore.Core/TransactionLog/Chunks/TFChunkDbConfig.cs
… src/EventStore.Core/TransactionLog/Chunks/TFChunkManager.cs
… src/EventStore.Core/TransactionLog/Chunks/TFChunkReader.cs
… src/EventStore.Core/TransactionLog/Chunks/TFChunkReaderExistsAtOptimizer.cs
… src/EventStore.Core/Util/Opts.cs
… src/EventStore.Core/VNodeBuilder.cs