# Problem/Solution description
PTable merges of index entries related to scavenged chunks are around 10x slower than merges of index entries related to non-scavenged chunks as shown below:

```
Merge with non-scavenged chunk:
[PID:30978:007 2018.01.08 09:50:48.759 TRACE PTable              ] PTables merge finished in 00:00:01.8044394 ([2048000, 2048000] entries merged into 4096000).
Merge with scavenged chunk:
[PID:32224:007 2018.01.08 10:13:22.025 TRACE PTable              ] PTables merge finished in 00:00:25.2660264 ([2048000, 2048000] entries merged into 4096000).
```
The slowness is due to the [existAt call](https://github.com/EventStore/EventStore/blob/release-v4.1.1/src/EventStore.Core/TransactionLog/Chunks/TFChunk/TFChunkReadSide.cs#L212) that verifies if an index entry exists in the chunk file by doing a binary search look up on the Posmap. The disk seeks done during the binary search add a significant amount of time.

The proposed solution is to add an option **--optimize-index-merge** that uses bloom filters to keep the posmap in memory for as many scavenged chunks as possible. This reduces the time for PTable merges and also decreases the disk pressure which is already high during PTable merges **(Full description of the solution in commit message)**

# Results:
## Test 1 (index rebuild of ~4M records on local machine/SSD with all chunks scavenged)
**Merge with scavenged chunk without optimization:**
 ```
PTables merge finished in 00:00:23.1940588 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:20.1107664 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:35.5952172 ([2000000, 2000000] entries merged into 4000000).
```
**Total merge time: 78s**

**Merge with scavenged chunk with --optimize-index-merge**
 ```
PTables merge finished in 00:00:03.1470243 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:03.1325550 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:04.3924874 ([2000000, 2000000] entries merged into 4000000).
```
**Total merge time: 10s (around 7-8x speedup)**

## Test 2 (index rebuild of ~20M records on AWS gp2/3000 iops from around 100 scavenged chunks)
**Merge with scavenged chunk without optimization:**
 ```
PTables merge finished in 00:00:26.1488981 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:26.3401015 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:54.1679085 ([2000000, 2000000] entries merged into 4000000).
PTables merge finished in 00:00:25.5928785 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:24.9906702 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:50.9753001 ([2000000, 2000000] entries merged into 4000000).
PTables merge finished in 00:01:53.2629083 ([4000000, 4000000] entries merged into 8000000).
PTables merge finished in 00:00:26.2846765 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:26.0847349 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:53.6072441 ([2000000, 2000000] entries merged into 4000000).
PTables merge finished in 00:00:26.0905945 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:26.4107351 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:54.8803166 ([2000000, 2000000] entries merged into 4000000).
PTables merge finished in 00:02:00.7227806 ([4000000, 4000000] entries merged into 8000000).
PTables merge finished in 00:03:56.1719077 ([8000000, 8000000] entries merged into 16000000).
PTables merge finished in 00:00:26.7429373 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:24.8559697 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:51.3573719 ([2000000, 2000000] entries merged into 4000000).
```
**Total merge time: ~986s**

**Merge with scavenged chunk with --optimize-index-merge**
 ```
PTables merge finished in 00:00:03.0629760 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:03.6487620 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:04.6183743 ([2000000, 2000000] entries merged into 4000000).
PTables merge finished in 00:00:02.8651277 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:03.2199336 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:04.9969473 ([2000000, 2000000] entries merged into 4000000).
PTables merge finished in 00:00:16.7856282 ([4000000, 4000000] entries merged into 8000000).
PTables merge finished in 00:00:03.1933536 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:02.8766726 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:04.6630766 ([2000000, 2000000] entries merged into 4000000).
PTables merge finished in 00:00:02.7591681 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:03.7337959 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:05.2891856 ([2000000, 2000000] entries merged into 4000000).
PTables merge finished in 00:00:16.9391175 ([4000000, 4000000] entries merged into 8000000).
PTables merge finished in 00:00:34.3026239 ([8000000, 8000000] entries merged into 16000000).
PTables merge finished in 00:00:03.1081373 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:03.0633281 ([1000000, 1000000] entries merged into 2000000).
PTables merge finished in 00:00:04.9363529 ([2000000, 2000000] entries merged into 4000000).
```
**Total merge time: ~122s (8x speedup)**

**For the optimized runs, the above tests also includes the time taken to create and populate the bloom filters which is done on the fly when calls are made to existAt.**

# Verification of data integrity
Verification of data integrity was done by:
- checking if the md5 hashes of the PTables and the checkpoints in `indexmap` match for the above 2 tests after the index rebuild
- doing index rebuilds with a few small --max-mem-table-size values (to generate many merges) with and without --optimize-index-merge and then checking if the md5 hashes of the final PTables match