@shaan1337 2018-01-30
Added additional test results based on 100 scavenged chunks to the PR details.
Speedup: 8x