@gregoryyoung 2018-02-08
What is happening on the outliers? I would expect the time to grow linearly but it seems to grow exponentially.