@gregoryyoung 2018-02-08
EG 

PTables merge finished in 00:00:05.2891856 ([2000000, 2000000] entries merged into 4000000).
PTables merge finished in 00:00:16.9391175 ([4000000, 4000000] entries merged into 8000000).
PTables merge finished in 00:00:34.3026239 ([8000000, 8000000] entries merged into 16000000)