@shaan1337 2018-02-09
The time for merging 2 1M PTable records also includes the time to create the bloom filters (afterwards in the test above these bloom filters are cached when merging 2M,4M,8M, etc.), so the time above for 1M records is a bit higher than is. If we ignore them:

2M -> 5s
4M -> 17s (~3.4x)
8M -> 34s (~2x)

It seems that the factor exponentially decreases and tends towards ~2x which is linear.
Not sure what it is but I think that there is a constant factor involved that's more significant when the times are close to 0.