@Lougarou 2018-04-09
### Another integrity check
ChunkSize set to 10 MB (in TFConsts.cs)
1. --max-mem-table-size=100 --optimize-index-merge --always-keep-scavenged
2. Write 10000 events to stream X
3. Truncate before stream X to 5000
4. Scavenge
5. Write 10000 events to stream Y - <random GUID>
6. Verify if all 5000 events are found in stream X correctly
For reference:
[bloomdataintegrity.zip](https://github.com/EventStore/EventStore/files/1889318/bloomdataintegrity.zip)